VIC-4717 Prevent eyes appearing behind time (#3931)
The show clock behavior demonstrates a corner case where we are
transitioning between on-screen eyes and non-eye display geometry,
like a clock face. In these cases there is no content in the
Procedural_Eyes layer since the eyes are just part of the png
sequence, but keepalive performances set up before this anim starts
are hanging around and inducing keepalive events resulting in eyes
displayed behind the intended content. At present, keepalives are only
updated if _streamingAnimation == nullptr anyway, so this change just
clears out any existing Keepalive keyframes in the
_proceduralTrackComponent any time we initialize a new anim for
streaming. If we eventually decide we want to have an animation run
with keepalive eyes in the background, this will need to be addressed
across the entire keepalive system.

* Make ProceduralClock use BaseStationTimer for internal transitions

BehaviorProceduralClock was using system time to determine how long
it should display the clock image for. Since its a delta, this has no
benefit I can think of over using BaseStationTimer, and it makes
debugging in simulation impossible. It now uses BaseStationTimer to
drive internal state transitions.

+++ engine/aiComponent/behaviorComponent/behaviors/timer/behaviorProceduralClock.h
@@ -42,7 +42,7 @@ public:
     _instanceParams.showClockCallback = callback;
   }
 
-  int GetTimeDisplayClock_sec() const { return _instanceParams.totalTimeDisplayClock_sec;}
+  float GetTimeDisplayClock_sec() const { return _instanceParams.totalTimeDisplayClock_sec;}
 
 
 protected:
@@ -78,7 +78,7 @@ protected:
   // Default behavior is to update the clock once per second for the length of display clock
   virtual void TransitionToShowClockInternal();
 
-  void SetTimeDisplayClock_sec(int displayTime_sec) { _instanceParams.totalTimeDisplayClock_sec = displayTime_sec;}
+  void SetTimeDisplayClock_sec(float displayTime_sec) { _instanceParams.totalTimeDisplayClock_sec = displayTime_sec;}
 
   // Function which builds and displays the proceduralClock - adds the 4 core digits on top
   // of any quadrant images passed into the function
@@ -101,7 +101,7 @@ private:
     AnimationTrigger getOutAnim;
     bool shouldTurnToFace = false;
     Json::Value layout;
-    int totalTimeDisplayClock_sec;
+    float totalTimeDisplayClock_sec;
     bool shouldPlayAudioOnClockUpdates = true;
 
     // Asset properties
@@ -115,7 +115,7 @@ private:
   };
 
   struct LifetimeParams{
-    int timeShowClockStarted = 0;
+    float timeShowClockStarted = 0.0f;
     BehaviorState currentState = BehaviorState::TurnToFace;
     SmartFaceID targetFaceID;
     bool hasBaseImageBeenSent = false;
